Squash personal details onto profile page
It doesn't really make sense to have a separate page for personal
details. By moving it onto the profile page, there is only two places
(profile and photo) that they need to go to to update information about
themselves.

The personal details page didn't support ajax or post-redirect-get, so
removing that page remove an instance of "are you sure you want to
resend the post data" prompt.