Add cache-busting on js
I can't for the life of me figure out how to get hostinger to change the
cache headers, so instead implement cache busting.

We compute the checksum of the file before we insert the reference to
it. This is inefficient, because we end up doing it for every resource
every request. However, we aren't too concerned with resource usage as
this is low traffic.